requirements.txt
altgraph==0.17.2
asgiref==3.4.1
bleach==4.1.0
camelcase==0.2
certifi==2021.10.8
chardet==4.0.0
colorama==0.4.4
Django==3.2.8
djangorestframework==3.14.0
python-ulid==1.1.0
markdown==3.4.1
django-filter==22.1
docutils==0.17.1
future==0.18.2
idna==3.2
keyring==23.2.1
multipledispatch==0.6.0
mysql-connector-python==8.0.26
psycopg2==2.9.5
packaging==21.0
pefile==2021.9.3
pip==21.2.4
pkginfo==1.7.1
Pygments==2.10.0
pyinstaller-hooks-contrib==2021.3
pyinstaller==4.5.1
pyparsing==2.4.7
pytz==2021.3
pywin32-ctypes==0.2.0
readme-renderer==30.0
requests-toolbelt==0.9.1
requests==2.26.0
rfc3986==1.5.0
setuptools==58.2.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
twine==3.4.2
urllib3==1.26.7
webencodings==0.5.1
wheel==0.37.0
gunicorn==20.1.0
djangorestframework-simplejwt==5.2.2